Updated serializer to ease the way of using
Registering Class now takes pointers to data members instead of MetaType enum.
New serializer allows to skip any data member that isnt supposed to be in save. In example you could have changed player gravity by changing values in save.dat. It no longer works like that.
The only down side is that the members of the class have to be public but hopefully that isnt a big problem.